Add siyi_sdk module
